# Optimization: cache results between TypeScript projects