--- a/src/CommandLineOptions.ts
+++ b/src/CommandLineOptions.ts
@@ -57,6 +57,8 @@ export function mainCommand(
     )
     .option('--output <path>', 'path to the output file', 'index.scip')
     .option('--progress-bar', 'whether to enable a rich progress bar')
+    .option('--no-progress-bar', 'whether to disable the rich progress bar')
+    .option('--global-caches', 'whether to disable global caches between TypeScript projects')
     .option('--no-global-caches', 'whether to disable global caches between TypeScript projects')
     .argument('[projects...]')
     .action((parsedProjects, parsedOptions) => {