@varungandhi-src 2022-10-17
Very nice! How did you recognize that this approach of caching the SourceFile values would help speed up type-checking?